requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
apischema==0.17.5
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
based58==0.1.1
cachetools==4.2.4
certifi==2025.4.26
charset-normalizer==3.4.2
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.3.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gspread==6.2.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.10
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
multidict==6.4.4
oauth2client==4.1.3
oauthlib==3.2.2
OSlash==0.6.3
playwright==1.52.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.3
pyserum==0.5.0a0
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
solana==0.25.0
solders==0.2.0
Telethon==1.40.0
types-cachetools==4.2.10
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
websockets==10.4
yarl==1.20.1
uvicorn

